+++ engines/tahi_standard_tasks/client/app/components/title-and-abstract-task.js
@@ -1,0 +1,12 @@
+import TaskComponent from 'tahi/pods/components/task-base/component';
+
+export default TaskComponent.extend({
+  paperNotEditable: Ember.computed.not('task.paper.editable'),
+  isNotEditable: Ember.computed.alias('task.completed'),
+
+  actions: {
+    focusOut() {
+      return this.get('task').save();
+    }
+  }
+});